implemented helper functions to return product data to be used throughout the app
